\name{Cterm}
\docType{data}
\alias{Cterm}
\title{data matrix}
\description{ 
     data matrix about terms x features
     1993 terms
     38 features
     topic is molecular biology , developmental biology
     jaccard radial base kernel is used
}
\usage{data(Cterm)}
\format{ a matrix of data }
\source{ a matrix of data }
\references{ a matrix of data }
\keyword{ datasets }
