\name{sk}
\docType{data}
\alias{sk}
\title{data matrix}
\description{ 
     data matrix about terms x features
     1993 terms
     38 features
     topic is molecular biology , developmental biology
     standard string kernel
}
\usage{data(sk)}
\format{ a matrix of data }
\source{ a matrix of data }
\references{ a matrix of data }
\keyword{ datasets }
