\name{term}
\docType{data}
\alias{term}
\title{data matrix about terms x features}
\description{ 
     data matrix about terms x features
     97 terms
     13 features
     topic is molecular biology , developmental biology
     jaccard radial base kernel is used
}
\usage{data(term)}
\format{ a matrix of data }
\source{ a matrix of data }
\references{ a matrix of data }
\keyword{ datasets }
